+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased](https://github.com/eps1lon/poe-mods/compare/v1.3.4..HEAD)
+### Fixed
+- Output path in `dev:schema` script
 
 ## [1.3.4](https://github.com/eps1lon/poe-mods/compare/v1.3.3...v1.3.4)  (2017-09-23)
 ### Added