+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/eps1lon/poe-mods/compare/v1.3.4..HEAD)
+## [Unreleased](https://github.com/eps1lon/poe-mods/compare/v1.3.5..HEAD)
+
+## [1.3.5](https://github.com/eps1lon/poe-mods/compare/v1.3.4..v1.3.5) (2017-09-25)
 ### Added
 - Orb Generators now consider the spawnweight when choosing a Mod
 - Maximum number of sockets for Items